// shared/app.enums.ts
export enum Gender {
	MALE = 'Male',
	FEMALE = 'Female',
	OTHER = 'Other'
}

export const ethnicities = [
	'Non-Hispanic White',
	'Black or African American',
	'Latino, Latina, or Hispanic American',
	'South Asian or Indian American',
	'East Asian or Asian American',
	'Middle Eastern or Arab American',
	'Native American or Alaskan Native',
	'Hawaiian or Pacific Islander',
	'Multiple ethnicity / Other',
	'Prefer not to answer'
];

export enum ClassYear {
	FRESHMAN = 'Freshman',
	SOPHOMORE = 'Sophomore',
	JUNIOR = 'Junior',
	SENIOR = 'Senior'
}

export const gradYears = [2019, 2020, 2021, 2022, 2023, 2024, 2025];

export enum Referral {
	SOCIAL_MEDIA = 'Social Media',
	WEBSITE = 'Website',
	FLYERS = 'Flyers',
	CLASS = 'Class Callout',
	FRIEND = 'Friend',
	OTHER = 'Other'
}

export enum ShirtSize {
	SMALL = 'S',
	MEDIUM = 'M',
	LARGE = 'L',
	EXTRA_LARGE = 'XL',
	EXTRA_EXTRA_LARGE = 'XXL'
}

export enum Status {
	PENDING = 'Pending',
	ACCEPTED = 'Accepted',
	REJECTED = 'Rejected',
	WAITLIST = 'Waitlist'
}

export enum Major {
	COMPUTER_SCIENCE = 'Computer Science',
	COMPUTER_ENGINEERING = 'Computer Engineering',
	ELECTRICAL_ENGINEERING = 'Electrical Engineering',
	MECHANICAL_ENGINEERING = 'Mechanical Engineering',
	ECONOMICS = 'Economics',
	BUSINESS = 'Business',
	SCIENCES = 'Sciences',
	LIBERAL_ARTS = 'Liberal Arts',
	FINE_ARTS = 'Fine Arts',
	UNDECIDED = 'Undecided',
	OTHER = 'Other'
}

export enum Resume {
	YES = 'Yes',
	NO = 'No'
}